+++ domutils/radar_tools/__init__.py
@@ -4,7 +4,9 @@ from .get_instantaneous     import get_instantaneous
 from .get_accumulation      import get_accumulation
 from .read_h5_composite     import read_h5_composite
 from .read_h5_vol           import read_h5_vol
+from .read_sqlite_vol       import read_sqlite_vol
 from .read_fst_composite    import read_fst_composite
 from .read_stage4_composite import read_stage4_composite
+from .read_mrms             import read_mrms
 from .plot_rdpr_rdqi        import plot_rdpr_rdqi
 from .                      import median_filter